chore: merge test ld_entities into ld_cb
